refactor: move Calibration/Maintenance into configs.py, remove prefix  on date
