+++ examples/fip_behavior_instrument.py
@@ -315,21 +315,21 @@ additional_device = Device(name="Photometry Clock")
 
 calibrations = [
     Calibration(
-        calibration_date=datetime(2023, 10, 2, 3, 15, 22, tzinfo=timezone.utc),
+        date=datetime(2023, 10, 2, 3, 15, 22, tzinfo=timezone.utc),
         device_name="470nm LED",
         description="LED calibration",
         input={"Power setting": [0]},
         output={"Power mW": [0.02]},
     ),
     Calibration(
-        calibration_date=datetime(2023, 10, 2, 3, 15, 22, tzinfo=timezone.utc),
+        date=datetime(2023, 10, 2, 3, 15, 22, tzinfo=timezone.utc),
         device_name="415nm LED",
         description="LED calibration",
         input={"Power setting": [0]},
         output={"Power mW": [0.02]},
     ),
     Calibration(
-        calibration_date=datetime(2023, 10, 2, 3, 15, 22, tzinfo=timezone.utc),
+        date=datetime(2023, 10, 2, 3, 15, 22, tzinfo=timezone.utc),
         device_name="560nm LED",
         description="LED calibration",
         input={"Power setting": [0]},
