+++ examples/fip_ophys_instrument.py
@@ -287,7 +287,7 @@ stimulus_device = d.RewardDelivery(
 additional_device = d.Device(name="Photometry Clock")
 
 calibration = d.Calibration(
-    calibration_date=datetime(2023, 10, 2, 3, 15, 22, tzinfo=timezone.utc),
+    date=datetime(2023, 10, 2, 3, 15, 22, tzinfo=timezone.utc),
     device_name="470nm LED",
     description="LED calibration",
     input={"Power setting": [1, 2, 3]},
