+++ src/aind_data_schema/components/devices.py
@@ -19,9 +19,8 @@ from aind_data_schema_models.units import (
 from pydantic import Field, ValidationInfo, field_validator, model_validator
 from typing_extensions import Annotated
 
-from aind_data_schema.base import GenericModel, GenericModelType, DataModel, AwareDatetimeWithDefault
+from aind_data_schema.base import GenericModelType, DataModel
 from aind_data_schema.components.coordinates import RelativePosition, Size3d
-from aind_data_schema.components.reagent import Reagent
 from aind_data_schema.components.identifiers import Software
 
 
@@ -274,28 +273,6 @@ class Device(DataModel):
     )
     port_index: Optional[str] = Field(default=None, title="Port index")
     additional_settings: Optional[GenericModelType] = Field(default=None, title="Additional parameters")
-    notes: Optional[str] = Field(default=None, title="Notes")
-
-
-class Calibration(DataModel):
-    """Generic calibration class"""
-
-    calibration_date: AwareDatetimeWithDefault = Field(..., title="Date and time of calibration")
-    device_name: str = Field(..., title="Device name", description="Must match a device name in rig/instrument")
-    description: str = Field(..., title="Description", description="Brief description of what is being calibrated")
-    input: GenericModelType = Field(GenericModel(), description="Calibration input", title="inputs")
-    output: GenericModelType = Field(GenericModel(), description="Calibration output", title="outputs")
-    notes: Optional[str] = Field(default=None, title="Notes")
-
-
-class Maintenance(DataModel):
-    """Generic maintenance class"""
-
-    maintenance_date: AwareDatetimeWithDefault = Field(..., title="Date and time of maintenance")
-    device_name: str = Field(..., title="Device name", description="Must match a device name in rig/instrument")
-    description: str = Field(..., title="Description", description="Description on maintenance procedure")
-    protocol_id: Optional[str] = Field(default=None, title="Protocol ID")
-    reagents: List[Reagent] = Field(default=[], title="Reagents")
     notes: Optional[str] = Field(default=None, title="Notes")
